Ny has to be positive

+++ include/Particles.h
@@ -65,7 +65,8 @@ public:
 
         for (int i=0; i<initial_num; i++)
         {
-            long double random=unidist(gen);
+//            long double random=unidist(gen);
+            double random=unidist(gen);
             int pos = random*Nxy;
             if(grid[pos]==0)
             {
@@ -142,7 +143,7 @@ public:
 
 
 //DIFFUSE PARTICLES
-    void attempt_diffusion(const int ind, long double &rand)
+    void attempt_diffusion(const int ind, double &rand)
     {
         // get diffuse particle
         diffuse_attempt++;
@@ -151,7 +152,8 @@ public:
         //choose random direction
         std::vector<Neighbour> n(lattice.get_neighbors2(particle_pos));
 
-        long double rand_size=rand*n.size();
+//        long double rand_size=rand*n.size();
+        double rand_size=rand*n.size();
         int dir=rand_size;
         rand = rand_size-dir;
         int new_pos = n[dir].position;
@@ -193,10 +195,10 @@ public:
 
 
 //BINDING ATTEMPT
-    void attempt_binding(double const alpha, double const J, int ind, long double &rand)
+    void attempt_binding(double const alpha, double const J, int ind, double &rand)
     {
         int pos = get_pos(ind);
-        long double rand_size = rand*3;
+        double rand_size = rand*3;
         int rand_int = rand_size;
         int ori = rand_int-1;
         rand=rand_size-rand_int;
@@ -279,7 +281,7 @@ public:
 
 
 //UNBINDING ATTEMPT
-    void attempt_unbinding(double const alpha, double const J, int ind,long double &rand)
+    void attempt_unbinding(double const alpha, double const J, int ind,double &rand)
     {
         // get bound particle
         unbinding_attempt++;
@@ -347,14 +349,15 @@ public:
 
         auto _is_allowed = [this,ori](const Neighbour &n){return is_interaction_allowed(ori,get_orientation(n.position),n.slope);};
 
-        auto _is_labelled = [this,&labels](const Neighbour &n){
-                    auto it= (std::find(begin(positions),end(positions),n.position));
+        auto _is_labelled = [this,&labels](const Neighbour &n)
+        {
+            auto it= (std::find(begin(positions),end(positions),n.position));
             if(it!= positions.end())
             {
                 int inds = it-positions.begin();
                 return labels[inds]!=0;
             }
-            };
+        };
 
 
         auto _label = [this,&labels,i,pos,&num_bonds](const Neighbour &n)
